Move web logic into a presentational layer
Resolves #292
