+++ src/knowlift/web/static/js/grade.js
